signin後にhomeにredirectする修正
home読み込み時にcostをgetするよう修正
