+++ src/SignUp.js
@@ -1,5 +1,21 @@
 import React from 'react'
 
+// TODO: SignUpの実装
+// handleSignUp(e) {
+//   e.preventDefault()
+
+//   const { email, password } = this.state;
+//   firebase.auth().createUserWithEmailAndPassword(email, password)
+//     .then((user) => {
+//       console.log(user);
+//       this.setState({ email: "", password: "" })
+//     })
+//     .catch((error) => {
+//       // Handle Errors here.
+//       var errorMessage = error.message;
+//       console.log(errorMessage);
+//     });
+// }
 class SignUp extends React.Component {
   render(){
     return(
